fix: UART begin/end pairing
Discovered bug where `HardwareSerial::begin()` is called multiple
times without calling `HardwareSerial::end()`. This manifested in
the firmware seizing up while writing logs. This only effects the
Swan, because of special logging that is enabled when a Swan is
the host MCU. The STM HAL function `uart_init` is not guaranteed
to be idempotent, therefore `uart_deinit` must be called between
successive initializations.

This commit also includes:
- better error handling for `note-c` calls
- logging of memory statistics behind the define `LOG_MEMORY_USAGE`